Use checked-in 7z to extract Inkscape due to COMPRESS-431
The 7z-gradle-plugin uses Commons Compress from Apache which does not
support multi input/output stream coders. See
https://issues.apache.org/jira/browse/COMPRESS-431 .
This issue occurs with both the Inkscape 7z archive and the official 7z
Extras archive, so it is not even possible to have that as an ivy
dependency to be extracted by the gradle plugin. And therefore, the 7z
Extras artifact has been re-archived as zip and the plugin replaced
entirely by an Exec task.